Do not treat all base images as unikraft images
Unikraft images use qemu as a platform and hence we need to define it
when wh use it a base. However, this is not the case for all OCI images.
Therefore, change the platform of the OCI image only when we use as a
base an image from Unikraft's catalog.

In the future, we might want to support a way to define if the OCI image
is Unikraft specific, so we can change the platform.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>