Fix and improve error reporting
When bunny acts as a frontend for buildkit, printing in stdout will
cause a fail. Instead, all errors should be reported as a return value
of the function we passs as an argument in the grpc client creation.

Moreover, this commit improves the error formating, but more work is
required.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>