Identify the type of file and call the respective parser
Since bunny wants to keep backwards compatibility with pun, we need to
keep the handling of Dockerfile-like files. Hence, we need to identofy
the format of the given file between a Dockerfile and a bunnyfile. For
that purpose, we check the first non-empty line of the file under
@syntax and if it starts with FROM we assume a Dockerfile and call the
respective function. Otherwise, we call the bunnyfile handler.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>