Create a Copy struct to allow a different source state than local
Currently, we were only able to copy files in the final image only from
the build context of the client. However, in the future, we want to be
able and copy files from other images or states. For that purpose, we
create a new struct to keep all the necessary information for a copy,
such as the source state, and the source, destination path.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>