Support the creation of initrd as a rootfs for Unikraft
Add support for building an initrd for Unikraft with the information in
the rootfs section of the bunnyfile. If the from field is empty or
scratch, then bunny will create a initrd with the files specified in the
include field of rootfs.

Furthermore, add checks regarding the input values of some fields in the
bunnyfile.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>