Update README and create examples for bunnyfile
Update the README file to provide information regarding the bunnyfile
and an explanation of each supported field for the current state.

In addition, move the annotations section in a separate file under the docs
directory.

Furthermore, create an examples directory and an example for Nginx and
Unikraft to further help the understanding and transition to bunny and
bunnyfile.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>