Refactor the way we transform hops to packInstructions
This commit signifficantly improves the readability and
extensibility of ToPack function. In particular, we create an interface
for unikernels to allow the easy support of new frameworks and use this
interface to create unikernel-specific files, such as rootfs.

Furthermore, we simplify the process of checking for reusing a prebuilt
rootfs from another image or locally.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>