Use mount instead of copies to get the cpio artifact
After the creation of the cpio file, we need to place this file in a
clean image that does not contain the tools we used for the creation of
the cpio. Therefore, instead of copying the file from one state to the
other, simply mount the final state that we will use under the ".boot/"
directory and hence when we write to this directory we immediately
writing to the final state/image that we will use.

Such design reduces the copies and the footprint overhead that comes
with them.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>