Fix copying kernel from local build context
There was a bug in the refactored code of bunny. In the case we need to
switch the base of the final image from the kernel to the rootfs, we
were using as the source of the kernel copy the final image base.
However, the final image base could be empty/scratch if the kernel needs
to get copied from the local build context.

Instead, we need to set up the copy of the kernel using the correct
LLB state and not the base of the final image.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>