Improve error handling in FilesLLB
THere is a case where the input of FIlesLLB is not formatted in the
correct way and hence an error should get returned. Although, we do
check for the format of the file list before we call the FilesLLB, it is
better to also include this check inside the funtion, since it is
exported through the hops package.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>