Update implementation of generic framework
Update the implementation of the generic implementation for frameworks
and in particular what is supported. The generic framework is supposed
to support all kind of types for rootfs and filesystem. The pupropse of
the generic implementation is to provide a quick and dirty solution for
the frameworks that have not been added to bunny yet. Therefore, and
since a framework is not officially supported, it is the responsibility
of the user to provide correct information.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>