Add support for appending environment variables
Add support for adding or appending environment variables in the final
container image. The users can use both bunnyfile and Containerfile-like
syntax files in order to set environment variables.

In the case of bunnyfile the environment variables are set under the
field `envs` as a list.

In the case of Containerfile-like syntax files, users can use the `ENV`
instruction as normally.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>