update Packinstructions struct adding more fields
Add a new struct (PackConfig) as a new field of Packinstructions to
store useful information for the configuration of the final image. For
the time being only the cmmonad line, the reference of the base image
and the monitor is stored, but in the future, we will store there,
environment variables, user/group and other configuration options of the
final image.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>