Refactor the way we annotate and set the config of the final image
Change the way we set up the configuration of the final image in order
to support the reuse of the base image's config. In that way, we can
maintain the env variables, labels and other information from the base
image without overwriting them.

Also, improve the readability of the code and create smaller helper
functions with a particular purpose.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>