Refactor, rewrite and break in smaller functions topack
Refactor the ToPack function to improve the readability and remove
complicated logic with complex nested if statements> Instead move some
of the logic in helper functions to assist with
performing all the necessary steps. In particular create the following
functions:
- handleKernel: Which is repsonsible to handle the kernel entry in
  bunnyfile
- handleRootfs: Which is repsonsible to handle the rootfs entry in
  bunnyfile
- makeCopy: which creates a new copy struct with the necessary
  information for a file copy from another state
- SetBaseAndGetPaths: Chooses the base of the final image based on the
  rootfs and kernel entries and returns the respective kernel and rootfs
  paths inside that base image.
- SetAnnotations: Sets up the necessary annotations in the
  PackInstructions struct.

Furthermore, create a new struct PackEntry, which maintains the
necessary information for a processed bunnyfile entry (i.e. rootfs, kernel)

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>